services/scene_analyzer_worker/requirements.txt
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
pydantic==2.5.0
